Can no longer remove own user.
As an added benefit to this, there won't ever be "no admin accounts" unless you log into the database and remove them all yourself.

+++ UserAdministration.php
@@ -67,7 +67,15 @@
     }
     if(isset($_POST["DeleteUser"]))
     {
-      DeleteUserConfirmation($_POST["UserIDForm"]); 
+      if ($_POST["UserIDForm"] == $_SESSION["UserID"])
+      {
+        $_SESSION["UserAdminEcho"] = "U kan niet uw eigen account verwijderen.";
+        header("Location: UserAdministration.php");
+      }
+      else
+      {
+        DeleteUserConfirmation($_POST["UserIDForm"]); 
+      }
     }
   ?>
 </div>
